api/requirements.txt
Django==5.2
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.7.0